// hairlab/src/app/models/enums/specialization.ts
export enum Specialization {

  WOMENS_CUT = 'WOMENS_CUT',

  MENS_CUT = 'MENS_CUT',

  PIXIE_CUT = 'PIXIE_CUT',

  BOB_CUT = 'BOB_CUT',

  CURLY_HAIR_CUT = 'CURLY_HAIR_CUT',

  BLOW_DRY = 'BLOW_DRY',

  STYLING = 'STYLING',

  UPDO = 'UPDO',

  HAIR_COLOR = 'HAIR_COLOR',

  BALAYAGE = 'BALAYAGE',

  HIGHLIGHTS = 'HIGHLIGHTS',

  COLOR_CORRECTION = 'COLOR_CORRECTION',

  CREATIVE_COLOR = 'CREATIVE_COLOR',

  BLEACHING = 'BLEACHING',

  HAIR_TREATMENTS = 'HAIR_TREATMENTS',

  SCALP_TREATMENTS = 'SCALP_TREATMENTS',

  BEARD_GROOMING = 'BEARD_GROOMING',

  SHAVING = 'SHAVING'
}